refactoring directories and names
