+++ main/java/AEPTools/AEPCataloguer.java
@@ -1,0 +1,11 @@
+package java.AEPTools;
+
+import java.ToolsInterfaces.CataloguerInterface;
+
+/**
+ * Basic AEPCataloguer that implements the AEPCataloguer Interface
+ */
+public class AEPCataloguer implements CataloguerInterface
+{
+
+}
