+++ main/java/AEPTools/AEPIngestor.java
@@ -1,0 +1,19 @@
+package java.AEPTools;
+
+import java.ParameterInterfaces.AuthTokenInterface;
+import java.ParameterInterfaces.DataSetIdInterface;
+import java.ParameterInterfaces.SchemaInterface;
+import java.ToolsInterfaces.IngestorInterface;
+import java.io.FileInputStream;
+
+/**
+ * Basic AEPIngestor that implements the IngestorInterface
+ */
+public class AEPIngestor implements IngestorInterface
+{
+
+    @Override
+    public void Upload(FileInputStream toUpload, SchemaInterface schema, DataSetIdInterface dataSetId, AuthTokenInterface authToken) {
+
+    }
+}
