--- main/java/Cataloguer.java
@@ -1,11 +1,0 @@
-package java;
-
-import java.Interfaces.CataloguerInterface;
-
-/**
- * Basic Cataloguer that implements the Cataloguer Interface
- */
-public class Cataloguer implements CataloguerInterface
-{
-
-}
