+++ main/java/Importer.java
@@ -1,6 +1,9 @@
 package java;
 
-import java.Interfaces.*;
+import java.AEPTools.AEPCataloguer;
+import java.AEPTools.AEPIngestor;
+import java.AEPTools.AEPValidator;
+import java.ToolsInterfaces.*;
 
 /**
  * Basic importer class that implements the ImporterInterface
@@ -13,9 +16,9 @@ public class Importer
     private ValidatorInterface schemaValidator;
 
     public Importer() {
-        dataCataloguer = new Cataloguer();
-        dataIngestor = new Ingestor();
-        schemaValidator = new Validator();
+        dataCataloguer = new AEPCataloguer();
+        dataIngestor = new AEPIngestor();
+        schemaValidator = new AEPValidator();
     }
 
 }
