+++ main/java/ToolsInterfaces/ValidatorInterface.java
@@ -1,0 +1,13 @@
+package java.ToolsInterfaces;
+
+import java.ParameterInterfaces.AuthTokenInterface;
+import java.ParameterInterfaces.DataSetIdInterface;
+import java.ParameterInterfaces.SchemaInterface;
+
+/**
+ * Interface to be used to create a Schema AEPValidator
+ */
+public interface ValidatorInterface {
+
+    public void validateSchema(SchemaInterface schema, DataSetIdInterface dataSetId, AuthTokenInterface authToken);
+}
